// src/zustand/interfaces/organisation-setting.interface.ts
export interface OrganisationSetting {
    id?: number;
    company_name: string;
    company_email: string;
    company_phone: string;
    company_address: string;
    facebook_url: string;
    youtube_url: string;
    tiktok_url: string;
    spotify_url: string;
    instagram_url: string;
    twitter_url: string;
    linkedin_url: string;
    logo?: string;
}